@@ -19,8 +19,8 @@
         "up-preview": "npm run package-clean && npm install --registry https://preview.registry.zenstack.dev -D zenstack@latest && npm install --registry https://preview.registry.zenstack.dev @zenstackhq/runtime@latest @zenstackhq/server@latest @zenstackhq/tanstack-query@latest"
     },
     "dependencies": {
+        "@auth/prisma-adapter": "^2.7.4",
         "@heroicons/react": "^2.0.12",
-        "@next-auth/prisma-adapter": "^1.0.5",
         "@prisma/client": "^5.19.1",
         "@tanstack/react-query": "^5.8.4",
         "@tanstack/react-query-devtools": "^5.8.3",
@@ -33,11 +33,11 @@
         "decimal.js": "^10.4.3",
         "moment": "^2.29.4",
         "nanoid": "^4.0.0",
-        "next": "^14.0.3",
-        "next-auth": "^4.24.5",
+        "next": "^15.0.3",
+        "next-auth": "^5.0.0-beta.25",
         "react": "^18.2.0",
         "react-dom": "^18.2.0",
-        "react-toastify": "^9.0.8",
+        "react-toastify": "^10.0.6",
         "superjson": "^1.12.0"
     },
     "devDependencies": {
@@ -48,13 +48,13 @@
         "@types/react-dom": "^18.2.0",
         "@typescript-eslint/eslint-plugin": "^6.13.1",
         "@typescript-eslint/parser": "^6.13.1",
-        "autoprefixer": "^10.4.12",
+        "autoprefixer": "^10.4.20",
         "eslint": "^7.32.0",
         "eslint-config-next": "12.3.1",
         "lower-case-first": "^2.0.2",
-        "postcss": "^8.4.16",
+        "postcss": "^8.4.49",
         "prisma": "^5.19.1",
-        "tailwindcss": "^3.1.8",
+        "tailwindcss": "^3.4.16",
         "typescript": "^5.3.2",
         "zenstack": "^2.9.4"
     }
