@@ -1,49 +1,42 @@
-/* eslint-disable @typescript-eslint/no-unsafe-member-access */
-/* eslint-disable @typescript-eslint/no-explicit-any */
-import { useCreateUser } from '@lib/hooks';
+'use client';
+
 import { signIn } from 'next-auth/react';
 import Image from 'next/image';
 import Link from 'next/link';
+import { useSearchParams } from 'next/navigation';
 import { FormEvent, useState } from 'react';
 import { toast } from 'react-toastify';
 
-export default function Signup() {
+export default function Signin() {
     const [email, setEmail] = useState('');
     const [password, setPassword] = useState('');
-    const signup = useCreateUser();
+    const params = useSearchParams();
 
-    async function onSignup(e: FormEvent<HTMLFormElement>) {
-        e.preventDefault();
-        try {
-            await signup.mutateAsync({ data: { email, password } });
-        } catch (err: any) {
-            console.error(err);
-            if (err.info?.prisma === true) {
-                if (err.info.code === 'P2002') {
-                    toast.error('User already exists');
-                } else {
-                    toast.error(`Unexpected Prisma error: ${err.info.code}`);
-                }
-            } else {
-                toast.error(`Error occurred: ${JSON.stringify(err)}`);
-            }
-            return;
+    const error = params.get('error');
+    if (error) {
+        if (error === 'OAuthAccountNotLinked') {
+            toast.error('Unable to signin. The user email may be already in use.');
+        } else {
+            toast.error(`Authentication error: ${error.toString()}`);
         }
+    }
 
+    async function onSignin(e: FormEvent<HTMLFormElement>) {
+        e.preventDefault();
         const signInResult = await signIn('credentials', {
             redirect: false,
             email,
             password,
         });
-        if (signInResult?.ok) {
+        if (!signInResult?.error) {
             window.location.href = '/';
         } else {
-            console.error('Signin failed:', signInResult?.error);
+            toast.error(`Signin failed. Please check your email and password.`);
         }
     }
 
     return (
-        <div className="flex flex-col items-center justify-center px-6 pt-4 lg:pt-8 mx-auto h-screen bg-cover bg-[url('/auth-bg.jpg')]">
+        <div className='flex flex-col items-center justify-center px-6 pt-4 lg:pt-8 mx-auto w-screen h-screen bg-cover bg-[url("/auth-bg.jpg")]'>
             <Link href="/">
                 <div className="flex space-x-4 items-center mb-6 lg:mb-10">
                     <Image src="/logo.png" width={42} height={42} alt="logo" />
@@ -52,8 +45,9 @@ export default function Signup() {
             </Link>
             <div className="items-center justify-center w-full bg-white rounded-lg shadow lg:flex md:mt-0 lg:max-w-screen-md xl:p-0">
                 <div className="w-full p-6 space-y-8 sm:p-8 lg:p-16">
-                    <h2 className="text-2xl font-bold text-gray-900 lg:text-3xl">Create a Free Account</h2>
-                    <form className="mt-8" action="#" onSubmit={(e) => void onSignup(e)}>
+                    <h2 className="text-2xl font-bold text-gray-900 lg:text-3xl">Sign in to your account</h2>
+
+                    <form className="mt-8" action="#" onSubmit={(e) => void onSignin(e)}>
                         <div className="mb-6">
                             <label htmlFor="email" className="block mb-2 text-sm font-medium text-gray-900">
                                 Your email
@@ -90,25 +84,32 @@ export default function Signup() {
                                     name="remember"
                                     type="checkbox"
                                     className="w-4 h-4 border-gray-300 rounded bg-gray-50 focus:ring-3 focus:ring-primary-300"
-                                    required
                                 />
                             </div>
                             <div className="ml-3 text-sm">
                                 <label htmlFor="remember" className="font-medium text-gray-900">
-                                    I accept the{' '}
-                                    <a href="#" className="text-primary-700 hover:underline">
-                                        Terms and Conditions
-                                    </a>
+                                    Remember me
                                 </label>
                             </div>
                         </div>
-                        <button className="btn btn-primary mt-4" type="submit">
-                            Create account
-                        </button>
+
+                        <div className="flex flex-col lg:flex-row gap-4 mt-4">
+                            <button className="btn btn-primary w-full lg:w-fit" type="submit">
+                                Login to your account
+                            </button>
+
+                            <div
+                                className="btn btn-outline w-full lg:w-fit"
+                                onClick={() => void signIn('github', { callbackUrl: '/' })}
+                            >
+                                Sign in with GitHub
+                            </div>
+                        </div>
+
                         <div className="mt-4 text-sm font-medium text-gray-500">
-                            Already have an account?{' '}
-                            <Link href="/signin" className="text-primary">
-                                Login here
+                            Not registered?{' '}
+                            <Link href="/signup" className="text-primary">
+                                Create account
                             </Link>
                         </div>
                     </form>
