@@ -1,24 +1,24 @@
+'use client';
+
 import { PlusIcon } from '@heroicons/react/24/outline';
-import { useCreateTodo, useFindManyTodo } from '@lib/hooks';
-import { List, Space } from '@prisma/client';
 import BreadCrumb from 'components/BreadCrumb';
 import TodoComponent from 'components/Todo';
 import WithNavBar from 'components/WithNavBar';
-import { GetServerSideProps } from 'next';
+import { useCurrentSpace } from 'lib/context';
+import { useCreateTodo, useFindManyTodo, useFindUniqueList } from 'lib/hooks';
+import { useParams } from 'next/navigation';
 import { ChangeEvent, KeyboardEvent, useState } from 'react';
-import { getEnhancedPrisma } from 'server/enhanced-db';
 
-type Props = {
-    space: Space;
-    list: List;
-};
-
-export default function TodoList(props: Props) {
+export default function TodoList() {
     const [title, setTitle] = useState('');
     const { mutate: createTodo } = useCreateTodo({ optimisticUpdate: true });
+    const params = useParams<{ listId: string }>();
+    const space = useCurrentSpace();
+
+    const { data: list } = useFindUniqueList({ where: { id: params.listId } });
 
     const { data } = useFindManyTodo({
-        where: { listId: props.list.id },
+        where: { listId: params.listId },
         include: {
             owner: true,
         },
@@ -35,22 +35,22 @@ export default function TodoList(props: Props) {
         createTodo({
             data: {
                 title,
-                list: { connect: { id: props.list.id } },
+                list: { connect: { id: params.listId } },
             },
         });
     };
 
-    if (!props.space || !props.list) {
+    if (!space || !list) {
         return <></>;
     }
 
     return (
         <WithNavBar>
             <div className="px-8 py-2">
-                <BreadCrumb space={props.space} list={props.list} />
+                <BreadCrumb space={space} list={list} />
             </div>
             <div className="container w-full flex flex-col items-center py-12 mx-auto">
-                <h1 className="text-2xl font-semibold mb-4">{props.list?.title}</h1>
+                <h1 className="text-2xl font-semibold mb-4">{list.title}</h1>
                 <div className="flex space-x-2">
                     <input
                         type="text"
@@ -79,28 +79,3 @@ export default function TodoList(props: Props) {
         </WithNavBar>
     );
 }
-
-export const getServerSideProps: GetServerSideProps<Props> = async ({ req, res, params }) => {
-    const db = await getEnhancedPrisma({ req, res });
-    const space = await db.space.findUnique({
-        where: { slug: params!.slug as string },
-    });
-    if (!space) {
-        return {
-            notFound: true,
-        };
-    }
-
-    const list = await db.list.findUnique({
-        where: { id: params!.listId as string },
-    });
-    if (!list) {
-        return {
-            notFound: true,
-        };
-    }
-
-    return {
-        props: { space, list },
-    };
-};
