@@ -1,6 +1,8 @@
+'use client';
+
 import { TrashIcon } from '@heroicons/react/24/outline';
-import { useDeleteTodo, useUpdateTodo } from '@lib/hooks';
 import { Todo, User } from '@prisma/client';
+import { useDeleteTodo, useUpdateTodo } from 'lib/hooks';
 import { ChangeEvent } from 'react';
 import Avatar from './Avatar';
 import TimeInfo from './TimeInfo';
