@@ -1,10 +1,14 @@
 /** @type {import('tailwindcss').Config} */
 module.exports = {
-    content: ['./pages/**/*.{js,ts,jsx,tsx}', './components/**/*.{js,ts,jsx,tsx}'],
+    content: [
+        './app/**/*.{js,ts,jsx,tsx,mdx}',
+        './pages/**/*.{js,ts,jsx,tsx,mdx}',
+        './components/**/*.{js,ts,jsx,tsx,mdx}',
+    ],
     theme: {
         extend: {},
     },
-    plugins: [require('daisyui'), require('@tailwindcss/line-clamp')],
+    plugins: [require('daisyui')],
     daisyui: {
         themes: ['light'],
     },
